backend/cubraos/teachify/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
black==23.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
flake8==6.1.0
frozenlist==1.8.0
idna==3.11
isort==5.13.2
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.1
Pillow==10.1.0
platformdirs==4.5.1
propcache==0.4.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.10.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.17.0
sqlparse==0.4.4
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
